+++ src/config/web.php
@@ -47,10 +47,12 @@ $config = [
         'formatter' => [
             'class' => 'yii\i18n\Formatter',
             'timeZone' => "Europe/Berlin",   // TODO: should be a user account attribute
-            'datetimeFormat' => 'php:Y-m-d H:i:00'
+            'datetimeFormat' => 'php:Y-m-d H:i:00',
+            'dateFormat' => 'php:Y-m-d'
         ],
         'session' => [
-            'class' => 'yii\web\DbSession',
+            'class'         => 'yii\web\DbSession',
+            'gCProbability' => 50    // 50% change to perform session GC
         ],        
     ],
     'params' => $params['app'],